S302T200-441 #time 45m Implemented unit tests for the deleteExpiredSaleListings method. Added empty list check to the deleteExpiredSaleListings method. Created new test class for SaleListingService so I could use mocking. ?test

+++ backend/src/main/java/org/seng302/project/service_layer/service/SaleListingService.java
@@ -764,7 +764,9 @@ public class SaleListingService {
         Specification<LikedSaleListing> spec = LikedSaleListingSpecification.saleListingClosesBefore(now);
         List<SaleListing> expiredListings = saleListingRepository.findAllByClosesBefore(now);
         List<LikedSaleListing> expiredLikedListings = likedSaleListingRepository.findAll(spec);
-        likedSaleListingRepository.deleteAll(expiredLikedListings);
+        if (!expiredLikedListings.isEmpty()) {
+            likedSaleListingRepository.deleteAll(expiredLikedListings);
+        }
         saleListingRepository.deleteAll(expiredListings);
     }
 }
